+++ src/utils/index.ts
@@ -11,3 +11,5 @@ export * from "./errors";
 export * from "./utf8-decoder";
 export * from "./message-processing";
 export * from "./sse";
+export * from "./model-validation";
+export * from "./streaming";